loadbalancers: lifting & shifting the resource ids
