+++ azurerm/internal/services/loadbalancer/parse/load_balancer_frontend_ip_configuration.go
