+++ azurerm/internal/services/loadbalancer/parse/load_balancer_inbound_nat_pool_test.go
