+++ azurerm/internal/services/loadbalancer/validate/load_balancer_inbound_nat_pool_id_test.go
