+++ azurerm/internal/services/loadbalancer/validate/load_balancer_inbound_nat_rule_id_test.go
